+++ tutorial/709_SLIM/check_mesh_for_issues.h
@@ -1,0 +1,3 @@
+#include <Eigen/Core>
+
+void check_mesh_for_issues(Eigen::MatrixXd& V, Eigen::MatrixXi& F);
