+++ tutorial/709_SLIM/get_soft_constraint_for_circle.h
@@ -1,0 +1,3 @@
+#include <Eigen/Core>
+
+void get_soft_constraint_for_circle(Eigen::MatrixXd& V_o, Eigen::MatrixXi& F, Eigen::VectorXi& b, Eigen::MatrixXd& bc);
